improve discovery of OPC UA client i) many servers do not return the endpoints populated with reachable Urls, so they need to be patched in GetEndpoints ii) if a FQDN is used to connect, the cert check always fails iii) if a cert doesn't contain a DC=, the cert domain check always fails
