+++ SampleApplications/SDK/Opc.Ua.Client/CoreClientUtils.cs
@@ -1,0 +1,176 @@
+/* ========================================================================
+ * Copyright (c) 2005-2013 The OPC Foundation, Inc. All rights reserved.
+ *
+ * OPC Foundation MIT License 1.00
+ * 
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without
+ * restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following
+ * conditions:
+ * 
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+ * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+ * OTHER DEALINGS IN THE SOFTWARE.
+ *
+ * The complete license agreement can be found here:
+ * http://opcfoundation.org/License/MIT/1.00/
+ * ======================================================================*/
+
+using System;
+using System.Collections.Generic;
+
+namespace Opc.Ua.Client
+{
+    /// <summary>
+    /// Defines numerous re-useable utility functions for clients.
+    /// </summary>
+    public class CoreClientUtils
+    {
+        #region Discovery
+        /// <summary>
+        /// Discovers the servers on the local machine.
+        /// </summary>
+        /// <param name="configuration">The configuration.</param>
+        /// <returns>A list of server urls.</returns>
+        public static IList<string> DiscoverServers(ApplicationConfiguration configuration)
+        {
+            List<string> serverUrls = new List<string>();
+
+            // set a short timeout because this is happening in the drop down event.
+            EndpointConfiguration endpointConfiguration = EndpointConfiguration.Create(configuration);
+            endpointConfiguration.OperationTimeout = 5000;
+
+            // Connect to the local discovery server and find the available servers.
+            using (DiscoveryClient client = DiscoveryClient.Create(new Uri("opc.tcp://localhost:4840"), endpointConfiguration))
+            {
+                ApplicationDescriptionCollection servers = client.FindServers(null);
+
+                // populate the drop down list with the discovery URLs for the available servers.
+                for (int ii = 0; ii < servers.Count; ii++)
+                {
+                    if (servers[ii].ApplicationType == ApplicationType.DiscoveryServer)
+                    {
+                        continue;
+                    }
+
+                    for (int jj = 0; jj < servers[ii].DiscoveryUrls.Count; jj++)
+                    {
+                        string discoveryUrl = servers[ii].DiscoveryUrls[jj];
+
+                        // Many servers will use the '/discovery' suffix for the discovery endpoint.
+                        // The URL without this prefix should be the base URL for the server. 
+                        if (discoveryUrl.EndsWith("/discovery"))
+                        {
+                            discoveryUrl = discoveryUrl.Substring(0, discoveryUrl.Length - "/discovery".Length);
+                        }
+
+                        // ensure duplicates do not get added.
+                        if (!serverUrls.Contains(discoveryUrl))
+                        {
+                            serverUrls.Add(discoveryUrl);
+                        }
+                    }
+                }
+            }
+
+            return serverUrls;
+        }
+
+        /// <summary>
+        /// Finds the endpoint that best matches the current settings.
+        /// </summary>
+        /// <param name="discoveryUrl">The discovery URL.</param>
+        /// <param name="useSecurity">if set to <c>true</c> select an endpoint that uses security.</param>
+        /// <returns>The best available endpoint.</returns>
+        public static EndpointDescription SelectEndpoint(string discoveryUrl, bool useSecurity, int operationTimeout = -1)
+        {
+            // needs to add the '/discovery' back onto non-UA TCP URLs.
+            if (discoveryUrl.StartsWith(Utils.UriSchemeHttps))
+            {
+                if (!discoveryUrl.EndsWith("/discovery"))
+                {
+                    discoveryUrl += "/discovery";
+                }
+            }
+
+            // parse the selected URL.
+            Uri uri = new Uri(discoveryUrl);
+
+            // set a short timeout because this is happening in the drop down event.
+            EndpointConfiguration configuration = EndpointConfiguration.Create();
+            if (operationTimeout > 0)
+            {
+                configuration.OperationTimeout = 5000;
+            }
+
+            EndpointDescription selectedEndpoint = null;
+
+            // Connect to the server's discovery endpoint and find the available configuration.
+            using (DiscoveryClient client = DiscoveryClient.Create(uri, configuration))
+            {
+                EndpointDescriptionCollection endpoints = client.GetEndpoints(null);
+
+                // select the best endpoint to use based on the selected URL and the UseSecurity checkbox. 
+                for (int ii = 0; ii < endpoints.Count; ii++)
+                {
+                    EndpointDescription endpoint = endpoints[ii];
+
+                    // check for a match on the URL scheme.
+                    if (endpoint.EndpointUrl.StartsWith(uri.Scheme))
+                    {
+                        // check if security was requested.
+                        if (useSecurity)
+                        {
+                            if (endpoint.SecurityMode == MessageSecurityMode.None)
+                            {
+                                continue;
+                            }
+                        }
+                        else
+                        {
+                            if (endpoint.SecurityMode != MessageSecurityMode.None)
+                            {
+                                continue;
+                            }
+                        }
+
+                        // pick the first available endpoint by default.
+                        if (selectedEndpoint == null)
+                        {
+                            selectedEndpoint = endpoint;
+                        }
+
+                        // The security level is a relative measure assigned by the server to the 
+                        // endpoints that it returns. Clients should always pick the highest level
+                        // unless they have a reason not too.
+                        if (endpoint.SecurityLevel > selectedEndpoint.SecurityLevel)
+                        {
+                            selectedEndpoint = endpoint;
+                        }
+                    }
+                }
+
+                // pick the first available endpoint by default.
+                if (selectedEndpoint == null && endpoints.Count > 0)
+                {
+                    selectedEndpoint = endpoints[0];
+                }
+            }
+
+            // return the selected endpoint.
+            return selectedEndpoint;
+        }
+        #endregion
+    }
+}
