+++ SampleApplications/Samples/ClientControls.Net4/Common/Client/ConnectServerCtrl.cs
@@ -56,6 +56,7 @@ namespace Opc.Ua.Client.Controls
         private ApplicationConfiguration m_configuration;
         private Session m_session;
         private int m_reconnectPeriod = 10;
+        private int m_discoverTimeout = 5000;
         private SessionReconnectHandler m_reconnectHandler;
         private CertificateValidationEventHandler m_CertificateValidation;
         private EventHandler m_ReconnectComplete;
@@ -285,7 +286,7 @@ namespace Opc.Ua.Client.Controls
             }
 
             // select the best endpoint.
-            EndpointDescription endpointDescription = ClientUtils.SelectEndpoint(serverUrl, UseSecurityCK.Checked);
+            EndpointDescription endpointDescription = CoreClientUtils.SelectEndpoint(serverUrl, UseSecurityCK.Checked, m_discoverTimeout);
 
             EndpointConfiguration endpointConfiguration = EndpointConfiguration.Create(m_configuration);
             ConfiguredEndpoint endpoint = new ConfiguredEndpoint(null, endpointDescription, endpointConfiguration);
@@ -398,7 +399,7 @@ namespace Opc.Ua.Client.Controls
                 }
 
                 // return the selected endpoint.
-                return ClientUtils.SelectEndpoint(discoveryUrl, UseSecurityCK.Checked);
+                return CoreClientUtils.SelectEndpoint(discoveryUrl, UseSecurityCK.Checked, m_discoverTimeout);
             }
             finally
             {
